api/requirements.txt
uvicorn
requests
fastapi
loguru
psycopg2-binary
SQLAlchemy
python-dotenv